scripts/requirements.txt
# Requirements for Zipline uploader script
requests>=2.25.0
python-dotenv>=1.0.0